backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# OCR
easyocr==1.7.1
Pillow==10.3.0

# ML
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# HTTP (for Ollama)
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
